break up graph_interact
