+++ src/components/dragelement/unhover.js
@@ -1,0 +1,43 @@
+/**
+* Copyright 2012-2016, Plotly, Inc.
+* All rights reserved.
+*
+* This source code is licensed under the MIT license found in the
+* LICENSE file in the root directory of this source tree.
+*/
+
+
+'use strict';
+
+
+var Events = require('../../lib/events');
+
+
+var unhover = module.exports = {};
+
+
+unhover.wrapped = function(gd, evt, subplot) {
+    if(typeof gd === 'string') gd = document.getElementById(gd);
+    // Important, clear any queued hovers
+    if(gd._hoverTimer) {
+        clearTimeout(gd._hoverTimer);
+        gd._hoverTimer = undefined;
+    }
+    unhover.raw(gd,evt,subplot);
+};
+
+
+// remove hover effects on mouse out, and emit unhover event
+unhover.raw = function unhover_raw(gd, evt) {
+    var fullLayout = gd._fullLayout;
+    if(!evt) evt = {};
+    if(evt.target &&
+       Events.triggerHandler(gd, 'plotly_beforehover', evt) === false) {
+        return;
+    }
+    fullLayout._hoverlayer.selectAll('g').remove();
+    if(evt.target && gd._hoverdata) {
+        gd.emit('plotly_unhover', {points: gd._hoverdata});
+    }
+    gd._hoverdata = undefined;
+};
