+++ src/components/modebar/buttons.js
@@ -11,6 +11,7 @@
 
 var Plotly = require('../../plotly');
 var Lib = require('../../lib');
+var setCursor = require('../../lib/setcursor');
 var Snapshot = require('../../snapshot');
 var Icons = require('../../../build/ploticon');
 
@@ -253,7 +254,7 @@ function handleCartesian(gd, ev) {
     Plotly.relayout(gd, aobj).then(function() {
         if(astr === 'dragmode') {
             if(fullLayout._hasCartesian) {
-                Plotly.Fx.setCursor(
+                setCursor(
                     fullLayout._paper.select('.nsewdrag'),
                     DRAGCURSORS[val]
                 );
